Improve security model
This introduces a Unix socket-based communication method for xremap in
GNOME, addressing limitations with the current DBUS implementation. This
change allows xremap to query the active window when running as a
separate user, improving flexibility and security.

It works with the Niri variant of xremap, which uses a very simple
socket protocol. Ideally xremap will adapt its GNOME variant to support
a socket and also gain some additional features:

- Implement a simpler socket protocol to fetch ActiveWindow rather than
  mimicking the Niri socket protocol.
- Re-check for socket existence to prevent permanent 'unsupported'
  state: GNOME (supported: false). The workaround for this is restarting
  the xremap process each time the socket file is created (can be done
  with systemd units).
- Delegate command execution to the GNOME extension instead of launching
  subprocesses directly.
- Improve multi-user support: Use a separate socket for each user so
  permissions can be tightened to prevent a user from hijacking another
  user's socket.